Refactor AdsEngine
